requirements.txt
streamlit
python-dotenv
langchain>=0.1.17
langchain-community
langchain-core
openai

duckdb
langchain-openai==0.1.7
faiss-cpu

# Document loaders
pypdf
docx2txt
unstructured
beautifulsoup4

# Tokenization
tiktoken

# Optional utilities
tqdm
watchdog